refactor: split standards_fetcher.py into focused modules
Extract network retry/integrity verification and ASVS JSON parsing into
separate internal modules. Update test patch targets accordingly.
